+++ weblate/checks/models.py
@@ -58,6 +58,7 @@ class WeblateChecksConf(AppConf):
         'weblate.checks.format.CSharpFormatCheck',
         'weblate.checks.format.JavaFormatCheck',
         'weblate.checks.format.JavaMessageFormatCheck',
+        "weblate.checks.format.PercentInterpolationCheck",
         "weblate.checks.format.I18NextInterpolationCheck",
         'weblate.checks.angularjs.AngularJSInterpolationCheck',
         'weblate.checks.qt.QtFormatCheck',