+++ weblate/settings_docker.py
@@ -838,6 +838,7 @@ CHECK_LIST = [
     "weblate.checks.format.CSharpFormatCheck",
     "weblate.checks.format.JavaFormatCheck",
     "weblate.checks.format.JavaMessageFormatCheck",
+    "weblate.checks.format.PercentInterpolationCheck",
     "weblate.checks.format.I18NextInterpolationCheck",
     "weblate.checks.angularjs.AngularJSInterpolationCheck",
     "weblate.checks.qt.QtFormatCheck",