+++ weblate/settings_example.py
@@ -664,6 +664,7 @@ CRISPY_TEMPLATE_PACK = "bootstrap3"
 #     'weblate.checks.format.CSharpFormatCheck',
 #     'weblate.checks.format.JavaFormatCheck',
 #     'weblate.checks.format.JavaMessageFormatCheck',
+#     "weblate.checks.format.PercentInterpolationCheck",
 #     "weblate.checks.format.I18NextInterpolationCheck",
 #     'weblate.checks.angularjs.AngularJSInterpolationCheck',
 #     'weblate.checks.qt.QtFormatCheck',